Refactor KeyStatistics component to use KeyStatisticsSkeleton for rendering. Introduce KeyStatisticsSkeleton component for improved structure and maintainability.
